requirements.txt
crawl4ai>=0.4.0
pinecone>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
playwright>=1.40.0
